+++ R/PredictionClassif.R
@@ -85,9 +85,8 @@
 #' @family Prediction
 #' @export
 #' @examples
-#' task = mlr_tasks$get("iris")
-#' learner = mlr_learners$get("classif.rpart")
-#' learner$predict_type = "prob"
+#' task = tsk("iris")
+#' learner = lrn("classif.rpart", predict_type = "prob")
 #' learner$train(task)
 #' p = learner$predict(task)
 #' p$predict_types