+++ R/PredictionRegr.R
@@ -45,8 +45,7 @@
 #' @export
 #' @examples
 #' task = mlr_tasks$get("boston_housing")
-#' learner = mlr_learners$get("regr.featureless")
-#' learner$predict_type = "se"
+#' learner = lrn("regr.featureless", predict_type = "se")
 #' p = learner$train(task)$predict(task)
 #' p$predict_types
 #' head(as.data.table(p))