+++ R/mlr_resamplings.R
@@ -25,19 +25,6 @@
 mlr_resamplings = R6Class("DictionaryResampling",
   inherit = Dictionary,
   cloneable = FALSE,
-
-  public = list(
-    get = function(key, ..., id = NULL, param_vals = NULL) {
-      obj = super$get(key, ...)
-      if (!is.null(id)) {
-        obj$id = id
-      }
-      if (!is.null(param_vals)) {
-        obj$param_set$values = insert_named(obj$param_set$values, param_vals)
-      }
-      obj
-    }
-  )
 )$new()
 
 #' @export