+++ tests/testthat/test_mlr_measures_selected_features.R
@@ -5,7 +5,7 @@ test_that("selected_features", {
   lrn = mlr_learners$get("classif.rpart")
   measures = mlr_measures$mget(c("classif.ce", "selected_features"))
 
-  rr = resample(task, lrn, mlr_resamplings$get("holdout"), ctrl = list(store_models = TRUE))
+  rr = resample(task, lrn, "holdout", ctrl = list(store_models = TRUE))
   perf = rr$aggregate(measures)
   expect_count(perf[["selected_features"]])
 })