+++ NEWS.md
@@ -1,5 +1,8 @@
 # mlr3 0.1.2
 
+* Refactored DataBackends.
+* Added functions to ease the creation of objects stored in dictionaries:
+  `tsk()`, `gen()`, `lrn()`, `rsp()`, `mea()`
 * New method `Task$rename()`.
 
 # mlr3 0.1.1